Restructure api_c files, extract handles api
