+++ libvcx/src/api_lib/api_c/mod.rs
@@ -1,0 +1,12 @@
+pub mod connection;
+pub mod credential;
+pub mod credential_def;
+pub mod disclosed_proof;
+pub mod issuer_credential;
+pub mod logger;
+pub mod proof;
+pub mod schema;
+pub mod utils;
+pub mod vcx;
+pub mod wallet;
+mod filters;
