+++ libvcx/src/api_lib/api_c/schema.rs
@@ -4,11 +4,12 @@ use indy_sys::CommandHandle;
 use libc::c_char;
 use serde_json;
 
-use crate::{schema, settings};
+use crate::api_lib::api_handle::schema;
+use crate::api_lib::utils_c::cstring::CStringUtils;
+use crate::api_lib::utils_c::runtime::execute;
 use crate::error::prelude::*;
-use crate::api_c::utils_c::cstring::CStringUtils;
+use crate::settings;
 use crate::utils::error;
-use crate::api_c::utils_c::runtime::execute;
 
 /// Create a new Schema object and publish corresponding record on the ledger
 ///
@@ -537,14 +538,14 @@ mod tests {
     #[allow(unused_imports)]
     use rand::Rng;
 
-    use crate::{api_c, libindy, settings, utils};
-    use crate::api_c::utils_c::return_types_u32;
-    use crate::schema::CreateSchema;
-    use crate::schema::tests::prepare_schema_data;
+    use crate::{api_lib, libindy, settings, utils};
+    use crate::api_lib::api_handle::schema::CreateSchema;
+    use crate::api_lib::api_handle::schema::tests::prepare_schema_data;
+    use crate::api_lib::utils_c::return_types_u32;
+    use crate::api_lib::utils_c::timeout::TimeoutUtils;
     #[allow(unused_imports)]
     use crate::utils::constants::{DEFAULT_SCHEMA_ATTRS, DEFAULT_SCHEMA_ID, DEFAULT_SCHEMA_NAME, SCHEMA_ID, SCHEMA_WITH_VERSION};
     use crate::utils::devsetup::*;
-    use crate::api_c::utils_c::timeout::TimeoutUtils;
 
     use super::*;
 
@@ -737,17 +738,17 @@ mod tests {
         {
             let cb = return_types_u32::Return_U32_U32::new().unwrap();
             let _rc = vcx_schema_get_state(cb.command_handle, handle, Some(cb.get_callback()));
-            assert_eq!(cb.receive(TimeoutUtils::some_medium()).unwrap(), api_c::PublicEntityStateType::Built as u32)
+            assert_eq!(cb.receive(TimeoutUtils::some_medium()).unwrap(), api_lib::PublicEntityStateType::Built as u32)
         }
         {
             let cb = return_types_u32::Return_U32_U32::new().unwrap();
             let _rc = vcx_schema_update_state(cb.command_handle, handle, Some(cb.get_callback()));
-            assert_eq!(cb.receive(TimeoutUtils::some_medium()).unwrap(), api_c::PublicEntityStateType::Published as u32);
+            assert_eq!(cb.receive(TimeoutUtils::some_medium()).unwrap(), api_lib::PublicEntityStateType::Published as u32);
         }
         {
             let cb = return_types_u32::Return_U32_U32::new().unwrap();
             let _rc = vcx_schema_get_state(cb.command_handle, handle, Some(cb.get_callback()));
-            assert_eq!(cb.receive(TimeoutUtils::some_medium()).unwrap(), api_c::PublicEntityStateType::Published as u32)
+            assert_eq!(cb.receive(TimeoutUtils::some_medium()).unwrap(), api_lib::PublicEntityStateType::Published as u32)
         }
     }
 }
