+++ libvcx/src/api_lib/api_handle/credential.rs
@@ -7,7 +7,7 @@ use crate::aries::{
     messages::a2a::A2AMessage,
     messages::issuance::credential_offer::CredentialOffer,
 };
-use crate::connection;
+use crate::api_lib::api_handle::connection;
 use crate::error::prelude::*;
 use crate::settings::indy_mocks_enabled;
 use crate::utils::constants::GET_MESSAGES_DECRYPTED_RESPONSE;
@@ -285,10 +285,10 @@ pub fn get_credential_status(handle: u32) -> VcxResult<u32> {
 
 #[cfg(test)]
 pub mod tests {
-    use crate::api_c::VcxStateType;
+    use crate::api_lib::VcxStateType;
     use crate::aries::messages::issuance::credential::Credential;
-    use crate::connection;
-    use crate::credential::{credential_create_with_offer, get_attributes, get_credential, send_credential_request};
+    use crate::api_lib::api_handle::connection;
+    use crate::api_lib::api_handle::credential::{credential_create_with_offer, get_attributes, get_credential, send_credential_request};
     use crate::error::VcxErrorKind;
     use crate::utils::devsetup::*;
     use crate::utils::mockdata::mockdata_credex::{ARIES_CREDENTIAL_OFFER, ARIES_CREDENTIAL_OFFER_JSON_FORMAT, ARIES_CREDENTIAL_RESPONSE, CREDENTIAL_SM_FINISHED, CREDENTIAL_SM_OFFER_RECEIVED};
