+++ libvcx/src/api_lib/api_handle/mod.rs
@@ -1,0 +1,7 @@
+pub mod connection;
+pub mod credential;
+pub mod credential_def;
+pub mod disclosed_proof;
+pub mod issuer_credential;
+pub mod proof;
+pub mod schema;
