+++ libvcx/src/api_lib/api_handle/schema.rs
@@ -3,7 +3,7 @@ use std::string::ToString;
 use serde_json;
 
 
-use crate::api_c::PublicEntityStateType;
+use crate::api_lib::PublicEntityStateType;
 use crate::error::prelude::*;
 use crate::libindy::utils::anoncreds;
 use crate::libindy::utils::ledger;
