+++ libvcx/src/api_lib/mod.rs
@@ -2,19 +2,8 @@ use std::fmt;
 
 #[macro_use]
 pub mod utils_c;
-
-pub mod vcx;
-pub mod connection;
-pub mod issuer_credential;
-pub mod proof;
-pub mod credential_def;
-pub mod schema;
-pub mod credential;
-pub mod disclosed_proof;
-pub mod wallet;
-pub mod logger;
-pub mod utils;
-mod filters;
+pub mod api_c;
+pub mod api_handle;
 
 
 /// This macro allows the VcxStateType to be
