+++ libvcx/src/api_lib/utils_c/ccallback.rs
