+++ libvcx/src/api_lib/utils_c/cstring.rs
