+++ libvcx/src/aries/handlers/connection/connection.rs
@@ -758,8 +758,9 @@ pub mod tests {
     use agency_client::mocking::AgencyMockDecrypted;
     use agency_client::update_message::{UIDsByConn, update_agency_messages};
 
-    use crate::{aries, connection, settings, utils};
-    use crate::api_c::VcxStateType;
+    use crate::{aries, settings, utils};
+    use crate::api_lib::api_handle::connection;
+    use crate::api_lib::VcxStateType;
     use crate::aries::messages::ack::tests::_ack;
     use crate::utils::constants;
     use crate::utils::devsetup::*;
