+++ libvcx/src/libindy/utils/wallet.rs
@@ -1,11 +1,11 @@
-use indy::future::Future;
 use indy::{ErrorCode, wallet};
 use indy::{INVALID_WALLET_HANDLE, SearchHandle, WalletHandle};
+use indy::future::Future;
 
 use crate::error::prelude::*;
 use crate::init::open_as_main_wallet;
-use crate::settings;
 use crate::libindy::utils::{anoncreds, signus};
+use crate::settings;
 use crate::settings::Actors::Issuer;
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
@@ -357,12 +357,13 @@ pub fn import(restore_config: &RestoreWalletConfigs) -> VcxResult<()> {
 pub mod tests {
     use agency_client::agency_settings;
 
+    use crate::api_lib;
+    use crate::api_lib::api_c;
     use crate::libindy::utils::signus::create_and_store_my_did;
     use crate::utils::devsetup::{SetupDefaults, SetupLibraryWallet, TempFile};
     use crate::utils::get_temp_dir_path;
 
     use super::*;
-    use crate::api_c;
 
     fn _record() -> (&'static str, &'static str, &'static str) {
         ("type1", "id1", "value1")
